+++ src/app/app.module.ts
@@ -9,14 +9,15 @@ import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 
 import { IonicStorageModule } from '@ionic/storage';
 
-import { StatusBar } from '@ionic-native/status-bar';
-import { SplashScreen } from '@ionic-native/splash-screen';
-
 import { StudentAccess } from './app';
 
 import { Store } from '../providers/store';
 import { Auth } from '../providers/auth';
 import { State } from '../providers/state';
+
+export function createTranslateLoader(http: Http) {
+    return new TranslateHttpLoader(http, './assets/i18n/', '.json');
+}
 
 @NgModule({
   declarations: [
@@ -34,9 +35,7 @@ import { State } from '../providers/state';
     TranslateModule.forRoot({
     loader: {
       provide: TranslateLoader,
-      useFactory: function loader(http: Http){
-        return new TranslateHttpLoader(http, './assets/i18n/', '.json');
-      },
+      useFactory: createTranslateLoader,
       deps: [ Http ]
     }
   })
@@ -47,11 +46,9 @@ import { State } from '../providers/state';
     Store,
     Auth,
     State,
-
-    StatusBar,
     HttpModule,
-    SplashScreen,
     {provide: ErrorHandler, useClass: IonicErrorHandler}
   ]
 })
-export class StudentAccessModule {}
+// renamed to AppModule to solve prod env bug ¯\_(ツ)_/¯
+export class AppModule {};