+++ src/components/components.module.ts
@@ -2,14 +2,17 @@ import { NgModule } from '@angular/core';
 import { IonicModule } from 'ionic-angular';
 
 import { Ribbon } from './ribbon/ribbon';
+import { GradeBadge } from './grade-badge/grade-badge';
 
 @NgModule({
   declarations: [
-    Ribbon
+    Ribbon,
+    GradeBadge
   ],
   imports: [ IonicModule ],
   exports: [
-    Ribbon
+    Ribbon,
+    GradeBadge
   ]
 })
 export class ComponentsModule {}