+++ src/pages/events/events.ts
@@ -1,5 +1,11 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import {
+  IonicPage,
+  NavController,
+  NavParams,
+  Loading,
+  LoadingController
+} from 'ionic-angular';
 
 import { Store } from '../../providers/store';
 
@@ -11,19 +17,26 @@ import { Store } from '../../providers/store';
 export class Events {
   public events;
   public selected;
+  private loading: Loading = this.loadingCtrl.create();
 
   constructor(
-    public nav: NavController,
-    public navParams: NavParams,
-    public store: Store
+    private nav: NavController,
+    private navParams: NavParams,
+    private loadingCtrl: LoadingController,
+    private store: Store
   ){}
 
-  ionViewDidLoad() {
+  async ionViewDidLoad(){
     this.selected = this.navParams.get('selected');
     if( !this.selected ){
-      this.store.get('EVENTS').then( (events = { events: [] }) => {
+      await this.loading.present();
+      try {
+        let events = await this.store.get('EVENTS');
         this.events = events.events;
-      });
+      } catch(err){
+        console.warn(err);
+      }
+      this.loading.dismiss();
     }
   }
   goSelected(item){